% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_setup.R
\name{end_log}
\alias{end_log}
\title{End Log
This function ends the log and closes the connection to the log file.}
\usage{
end_log(log_info)
}
\arguments{
\item{log_info}{This is the variable that stores the start_log call.}
}
\description{
End Log
This function ends the log and closes the connection to the log file.
}
