% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_setup.R
\name{getname}
\alias{getname}
\title{This function gets the name of the script being used in order to log that information}
\usage{
getname()
}
\value{
A string with the name of the script
}
\description{
This function gets the name of the script being used in order to log that information
}
