% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\docType{data}
\name{inputLogger}
\alias{inputLogger}
\title{Input logger}
\format{
An object of class \code{Logger} (inherits from \code{R6}) of length 13.
}
\usage{
inputLogger
}
\description{
Default logger used to log inputs in \code{\link[=logApp]{logApp()}}.
}
\keyword{datasets}
