% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{is.log}
\alias{is.log}
\title{Log Check}
\usage{
is.log(obj)
}
\arguments{
\item{obj}{Object to check.}
}
\value{
\code{TRUE} if object is a logger,
and \code{FALSE} otherwise.
}
\description{
Log Check
}
\examples{
info <- Logger$new("INFO")
is.log(info)

}
