% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{logApp}
\alias{logApp}
\title{Default Logger for Shiny.}
\usage{
logApp(ui, server, ..., inputs = logInputs)
}
\arguments{
\item{ui, server}{UI definition and server function
as passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{...}{Any other arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{inputs}{Logger to log inputs, set to \code{NULL} to not log.}
}
\description{
Serve a shiny application with default loggers,
useful for debugging.
}
