% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\docType{data}
\name{logInputs}
\alias{logInputs}
\title{Input logger}
\format{
An object of class \code{Logger} (inherits from \code{R6}) of length 12.
}
\usage{
logInputs
}
\description{
Default logger used to log inputs in \code{\link[=logApp]{logApp()}}.
}
\keyword{datasets}
