% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\name{plumber}
\alias{plumber}
\alias{prLog}
\alias{prWithLog}
\title{Plumber}
\usage{
prLog(file = NULL, ..., requests = requestLogger)

prWithLog(pr, requests = requestLogger)
}
\arguments{
\item{file}{Plumber file as passed to \code{\link[plumber:pr]{plumber::pr()}}.}

\item{...}{Any other argument to pass to \code{\link[plumber:pr]{plumber::pr()}}.}

\item{requests}{Logger to log requests, set to \code{NULL} to not log.}

\item{pr}{Plumber API as returned by \code{\link[plumber:pr]{plumber::pr()}}.}
}
\description{
Serve a plumber API with metrics, use this function
like you would use \code{\link[plumber:pr]{plumber::pr()}}.
}
