% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_setup.R
\name{start_log}
\alias{start_log}
\title{Start Logging}
\usage{
start_log(log_directory, script_name = getname())
}
\arguments{
\item{log_directory}{Path to the folder for the log file}

\item{script_name}{Name of the script that is being logged}
}
\description{
This function creates a log file that records messages and outputs from your script or markdown, the time it started running and the time it finished, and records the session info.
Note: Store this variable so that we can end it using the end_log function. I suggest naming it log or log_info.
}
