% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{templates}
\alias{templates}
\alias{infoLog}
\alias{errorLog}
\alias{warnLog}
\alias{successLog}
\title{Template Loggers}
\usage{
infoLog(prefix = "INFO")

errorLog(prefix = "ERROR")

warnLog(prefix = "WARNING")

successLog(prefix = "SUCCESS")
}
\arguments{
\item{prefix}{The prefix to use, this is passed to \link{Logger}.}
}
\description{
Template loggers for convenience.
}
