% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{Logger}
\alias{Logger}
\title{Logger}
\description{
Create a logger
}
\examples{
infoLog <- Logger$new("INFO")$
 date()$
 time()$
 hook(crayon::blue)

infoLog$log("Hello")
Sys.sleep(.7)
infoLog$log("World")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Logger$new()}}
\item \href{#method-date}{\code{Logger$date()}}
\item \href{#method-time}{\code{Logger$time()}}
\item \href{#method-unix}{\code{Logger$unix()}}
\item \href{#method-wd}{\code{Logger$wd()}}
\item \href{#method-hook}{\code{Logger$hook()}}
\item \href{#method-dir}{\code{Logger$dir()}}
\item \href{#method-flag}{\code{Logger$flag()}}
\item \href{#method-log}{\code{Logger$log()}}
\item \href{#method-dump}{\code{Logger$dump()}}
\item \href{#method-clone}{\code{Logger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$new(prefix = "", write = FALSE, file = "log.log", sep = "\\t")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{String to prefix all log messages.}

\item{\code{write}}{Whether to write the log to the \code{file}.}

\item{\code{file}}{Name of the file to dump the logs to,
only used if \code{write} is \code{TRUE}.}

\item{\code{sep}}{Separator between \code{prefix} and other flags
and messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-date"></a>}}
\if{latex}{\out{\hypertarget{method-date}{}}}
\subsection{Method \code{date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$date(format = "\%d-\%m-\%Y")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{Formatter for the item, passed
to \code{\link[=format]{format()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Include the date in the log
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-time"></a>}}
\if{latex}{\out{\hypertarget{method-time}{}}}
\subsection{Method \code{time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$time(format = "\%H:\%M:\%S")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{Formatter for the item, passed
to \code{\link[=format]{format()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Include the time in the log
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unix"></a>}}
\if{latex}{\out{\hypertarget{method-unix}{}}}
\subsection{Method \code{unix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$unix()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Include the time in the log
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wd"></a>}}
\if{latex}{\out{\hypertarget{method-wd}{}}}
\subsection{Method \code{wd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$wd()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Include the working directory
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hook"></a>}}
\if{latex}{\out{\hypertarget{method-hook}{}}}
\subsection{Method \code{hook()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$hook(fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{A function that accepts one argument (string)
and returns a modified version of that string.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Preprocess the prefix with a custom function
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dir"></a>}}
\if{latex}{\out{\hypertarget{method-dir}{}}}
\subsection{Method \code{dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$dir()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Include the directory in the log
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-flag"></a>}}
\if{latex}{\out{\hypertarget{method-flag}{}}}
\subsection{Method \code{flag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$flag(what)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{Function to run for every message logged
or string to include in log message.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Pass a custom flag
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$log(msg = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{Message to log}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Log messages
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dump"></a>}}
\if{latex}{\out{\hypertarget{method-dump}{}}}
\subsection{Method \code{dump()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$dump(file = "dump.log")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Name of the file to dump the logs to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Dump the log to a file
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
